+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Current
 
 - change semantics of Cuz_cohttp.accept_gzde and make it more flexible
+- rename Cuz_cohttp.accept_gzde to Cuz_cohttp.update_header
+- use ptime instead of Unix.gettimeofteday for the gzip configuration
 
 # 0.0.1